linux-user: Execve syscalls can use the qemu interpreter
If the interpreter is started with -execve as an argument,
on startup, qemu will attempt to find the path to itself
using getauxval() or readlink(/proc/self/exe).

Any calls to execve() that qemu tries to handle will be
run in a qemu interpreter process.

This allows using qemu-user inside a chroot'd rootfs, with
use of all binaries available as if it was native.